2023.09.16-06:45

+++ misc/starbed/scripts/login.py
@@ -11,6 +11,7 @@ parser.add_argument("node", help="specify network-node name")
 parser.add_argument("-c", "--command", default="bash")
 parser.add_argument("-i", "--inventory", default="hosts.small.yaml")
 parser.add_argument("-H", "--host", action='store_true')
+parser.add_argument("-S", "--ssh", action='store_true')
 args = parser.parse_args()
 
 # Load inventory file
@@ -25,13 +26,17 @@ if host is None:
   sys.exit("Error")
 
 # Execute remote login command
+if args.ssh:
+  cmd = f"ssh -t {host} {args.command}"
+  subprocess.run(cmd, shell=True)
+  sys.exit(0)
 if args.host:
   cmd = f"docker -H ssh://{host} run --rm "
   cmd += f"--net host -it -e PS1='{args.node}-{host}> ' "
   cmd += f"nicolaka/netshoot {args.command}"
   subprocess.run(cmd, shell=True)
-else:
-  cmd = f"docker -H ssh://{host} run --rm "
-  cmd += f"--net container:{args.node} -it -e PS1='{args.node}> ' "
-  cmd += f"nicolaka/netshoot {args.command}"
-  subprocess.run(cmd, shell=True)
+  sys.exit(0)
+cmd = f"docker -H ssh://{host} run --rm "
+cmd += f"--net container:{args.node} -it -e PS1='{args.node}> ' "
+cmd += f"nicolaka/netshoot {args.command}"
+subprocess.run(cmd, shell=True)
